fastApi_BackEnd/requirements.txt
fastapi[all]
pyjwt
uvicorn
pydantic
requests
httpx
aio-pika
openai
chromadb
langchain[all]
langchain_community
langchain-openai
authlib
python-dotenv